Update patch set 6
Patch Set 6:

I think we should submit as is, but I should mention that aosp/973584 (which is essentially the equivalent CL for RV) broke GMM when RV was migrated to Google3 (partially because GMM was using an older NestedScrollingParent interface implementation).  So, this CL may break Google3, but we can cross that bridge if we must.

Patch-set: 6
Reviewer: Gerrit User 1344416 <1344416@85c56323-6fa9-3386-8a01-6480fb634889>